huge commit! read whole commitMessage for details
User can create BuyOrder that will be displayed and deletable from own profileView.
User can't create BuyOrders that are already satisfiable! If so, the user shall use search and buy item by himself
A BuyOrder has hardcoded quantity of 1!
A BuyOrder will be deleted if successful.
A BuyOrder buys as soon as a item is ready to buy (= if item is activated, the item is ready to be bought)
You can't see other Users BuyOrders.
If multiple BuyOrders want to buy the same item the first BuyOrder that gets the opportunity will buy it. This shoud correspond to the CreationTime(not sure, maybe just random)!