implemented a new ItemMergingSystem, for more details read commit message
Item with similar Name, Price and Owner can be merged be Owner.
The MergingButton is showed in the own profileView.
If there are multiple similar Items, the View have multiple Buttons to merge.
In the MergeView the two Items are Listed.
The slaveItem (second) will be merged into the MasterItem (first).
The slaveItem will be deleted.
The images of the slaveItem will be deleted and the MasterItem keeps its previous images
After Merge the MasterItems quantity will be MasterItem.quantity + SlaveItem.quantity